refactor: move projections and rename GET slices
