+++ src/aggregates/case/index.ts
@@ -4,7 +4,7 @@ import type { Aggregate } from '../../shared/aggregate.js';
 import { registerCreateCaseRoutes } from './create-case/http.js';
 import { registerAddInteractionRoutes } from './add-interaction/http.js';
 import { registerCloseCaseRoutes } from './close-case/http.js';
-import { registerProjectCaseRoutes } from './project-case/http.js';
+import { registerGetCaseRoutes } from './get-case/http.js';
 import { registerOpenCasesRoutes } from './open-cases/http.js';
 
 export class CaseAggregate implements Aggregate {
@@ -12,7 +12,7 @@ export class CaseAggregate implements Aggregate {
     registerCreateCaseRoutes(router, eventStore);
     registerAddInteractionRoutes(router, eventStore);
     registerCloseCaseRoutes(router, eventStore);
-    registerProjectCaseRoutes(router, eventStore);
+    registerGetCaseRoutes(router, eventStore);
     registerOpenCasesRoutes(router, eventStore);
   }
 }
