+++ src/aggregates/client/get-client/http.ts
@@ -1,10 +1,9 @@
 import { Router } from 'express';
-import { projectContact } from './index.js';
+import { projectClient } from '../project-client.js';
 import { createTraceContext } from '../../../shared/trace.js';
 import type { EventStore } from '../../../shared/event-store.js';
 
-export function registerProjectContactRoutes(router: Router, eventStore: EventStore) {
-  // Reuse same helper as others
+export function registerGetClientRoutes(router: Router, eventStore: EventStore) {
   function extractTraceFromHeaders(headers: Record<string, unknown>) {
     return createTraceContext({
       traceId: headers['x-trace-id']?.toString(),
@@ -14,24 +13,24 @@ export function registerProjectContactRoutes(router: Router, eventStore: EventSt
     });
   }
 
-  router.get('/contacts/:id', async (req, res) => {
+  router.get('/clients/:id', async (req, res) => {
     const trace = extractTraceFromHeaders(req.headers);
-    const contactId = req.params.id;
+    const clientId = req.params.id;
     const startTime = Date.now();
 
     try {
-      const events = await eventStore.getEventsForAggregate('contact', contactId);
-      const state = projectContact(events);
+      const events = await eventStore.getEventsForAggregate('client', clientId);
+      const state = projectClient(events);
 
       if (!state) {
-        return res.status(404).json({ error: 'Contact not found' });
+        return res.status(404).json({ error: 'Client not found' });
       }
 
       const durationMs = Date.now() - startTime;
-      console.log(`[ContactFetched]`, { traceId: trace.traceId, spanId: trace.spanId, contactId, durationMs });
+      console.log(`[ClientFetched]`, { traceId: trace.traceId, spanId: trace.spanId, clientId, durationMs });
       return res.status(200).json(state);
     } catch (err) {
-      console.error('[get-contact error]', err);
+      console.error('[get-client error]', err);
       return res.status(500).json({ error: 'Internal server error' });
     }
   });
