+++ src/aggregates/contact/get-contact/http.ts
@@ -1,9 +1,10 @@
 import { Router } from 'express';
-import { projectCase } from './index.js';
+import { projectContact } from '../project-contact.js';
 import { createTraceContext } from '../../../shared/trace.js';
 import type { EventStore } from '../../../shared/event-store.js';
 
-export function registerProjectCaseRoutes(router: Router, eventStore: EventStore) {
+export function registerGetContactRoutes(router: Router, eventStore: EventStore) {
+  // Reuse same helper as others
   function extractTraceFromHeaders(headers: Record<string, unknown>) {
     return createTraceContext({
       traceId: headers['x-trace-id']?.toString(),
@@ -13,25 +14,26 @@ export function registerProjectCaseRoutes(router: Router, eventStore: EventStore
     });
   }
 
-  router.get('/cases/:id', async (req, res) => {
+  router.get('/contacts/:id', async (req, res) => {
     const trace = extractTraceFromHeaders(req.headers);
-    const caseId = req.params.id;
+    const contactId = req.params.id;
     const startTime = Date.now();
 
     try {
-      const events = await eventStore.getEventsForAggregate('case', caseId);
-      const state = projectCase(events);
+      const events = await eventStore.getEventsForAggregate('contact', contactId);
+      const state = projectContact(events);
 
       if (!state) {
-        return res.status(404).json({ error: 'Case not found' });
+        return res.status(404).json({ error: 'Contact not found' });
       }
 
       const durationMs = Date.now() - startTime;
-      console.log(`[CaseFetched]`, { traceId: trace.traceId, spanId: trace.spanId, caseId, durationMs });
+      console.log(`[ContactFetched]`, { traceId: trace.traceId, spanId: trace.spanId, contactId, durationMs });
       return res.status(200).json(state);
     } catch (err) {
-      console.error('[get-case error]', err);
+      console.error('[get-contact error]', err);
       return res.status(500).json({ error: 'Internal server error' });
     }
   });
 }
+
